Cleanup of ALU components. Creating common abstract superclass
