+++ src/test/java/org/karnes/homebrew/emulator/component/ArithemeticUnitTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
 import org.karnes.homebrew.bitset.FixedBitSet;
 import org.karnes.homebrew.emulator.ConditionCode;
-import org.karnes.homebrew.emulator.component.arithmeticunit.AU_OPCODE;
-import org.karnes.homebrew.emulator.component.arithmeticunit.ArithmeticUnit;
+import org.karnes.homebrew.emulator.component.alu.arithmeticunit.AU_OPCODE;
+import org.karnes.homebrew.emulator.component.alu.arithmeticunit.ArithmeticUnit;
 import org.karnes.homebrew.emulator.component.bus.BidirectionalBus;
 import org.karnes.homebrew.emulator.component.bus.VirtualBus;
 import org.karnes.homebrew.emulator.component.bus.connection.ReadableBusConnection;
